requirements.txt
requests==2.31.0
python-telegram-bot==21.0.1
python-dotenv==1.0.0
openai==0.28.1 